Introduce the `ProjectStackManager` interface
This commit introduces `ProjectStackManager`, an interface for
abstracting the persistence of `ProjectStack`s. Initially we thread this
through the `config get` command, picking one of two implementations
depending on the `--config-file` flag's absence or presence:

* If the flag is absent, the `DefaultProjectStackManager` wraps the
  existing "detection" path for finding configuration based on the
  project.
* If the flag is present, the `SpecifiedProjectStackManager` wraps the
  "explicit" path where the passed path is used to retrieve stack
  configuration.

As part of this work, the `ConfigFile` global is passed down to the
`get` command, where it is used to pick the right instance. Once we've
completed this refactoring for all `config ...` commands, we'll be able
to remove those uses of the global. A similar refactoring will let us do
this for other commands that use `cmdStack.ConfigFile` later on.