Warn if `refresh` or `destroy` use older plugins (#12196)
`refresh` and `destroy` operations do not run the user program, and
instead rely on the plugin versions specified in the state. This can
lead to confusing behavior where a user has updated a provider
dependency in their program, but these operations ignore that setting.
This change prints a warning if either of these operations will use an
older plugin version than the one specified in the program dependencies.

> [!NOTE]  
> This may be redundant once a fix for #12177 lands.

Co-authored-by: Will Jones <will@sacharissa.co.uk>