Update the `refresh`/`destroy` warnings for parameterized packages
This updates the warnings added by #12196 to also warn about updates to
parameterized packages, not just plugins.

Before this the warning would kick in if the underlying parameterised
plugin was updated (e.g. terraform-provider went from 0.3.0 to 0.3.1)
but it wouldn't warn if the user had regenerated the parameterised
package (e.g hashicorp/random 3.6.0 to hashicorp/random 3.6.3). This
change fixes it so the later will now print the expected warning.

It also fixes it so the warning refers to the package name, not the
plugin name (i.e. in the above terraform random example we print 'random
3.6.3' not 'terraform-provider 0.3.1').

Part of #17507