Factor out conformance testing framework (#17999)
This commit moves the core of the conformance testing framework into a
`tests` subpackage. From here, the plan would be to split out e.g.
`tests/l1_empty.go` etc. to factor out the tests themselves, since
`tests.go` and the `languageTests` array is getting quite big at this
point. Having the tests factored a bit better might also open the door
for discussing per-test providers and the like (though this may not be
feasible).