Add ConfigFile to GetConfig and SetConfig Ops In Go Automation API (#17939)
Add ConfigFile to ConfigOptions for GetConfig and SetConfig operations
to allow for users to specify a config file to read and write to during
these operations.

Add a new GetAllConfigWithOptions function to allow users to specify a
config file to read from and also if they want to show secrets

Error handling has been added to operations that do not support specific
config options such as SetAllConfigWithOptions returning a newAutoError
if ShowSecrets is set to true in the ConfigOptions.

GetAllConfig has been refactored to use the new GetAllConfigWithOptions
function to follow the pattern the other functions are using.

Test cases have been added to test the new functionality and error
handling and also to verify that the GetallConfig refactor works as it
previously did.