Add winget install test back to download cron workflow (#16980)
This was originally included in the download cron workflow, but had to
be removed because there wasn't a reliable way to use the winget CLI
from GitHub's runners. There's now a GitHub Action available for
installing the winget CLI, which we should be able to use now for this.

I temporarily set up the workflow to run on PRs (now removed). You can
see a successful run here:
https://github.com/pulumi/pulumi/actions/runs/10407763146/job/28823717019?pr=16980

Fixes #11342