Avoid unnecessary Node.js installations
When running `pulumi install --use-language-version-tools`, and `fnm` is installed on the system, we will automatically install the Node.js version specified in `.nvmrc` or `.node-version`.

When the file does not specify the full version (as recommended), for example `22`, we would always install the latest release of Node.js 22, even if a releasse was already present on the system. This is unnecessary, and we should use the available install.

This mostly affects Pulumi Deployments, where this installation flag is used by default. Whenever we release a new version of Pulumi, we also release new versions of the docker containers. These containers include the latest version of Node.js (22, and other versions) at the time of release, so the pre-installed version is usually very recent. However if a new Node.js version is released after our release, deployments that specify a Node.js version will always install that.

By using `fnm use ${VERSION} —install-if-missing`, we prefer the pre-installed version, and only install Node.js versions if the fnm version matching can not be satisfied by any of the pre-installed versions.