Load all schema over gRPC in conformance tests (#18043)
During the execution of a normal Pulumi program, the various components
(engine, language hosts, and providers) communicate over gRPC. When
running conformance tests, we want to simulate this environment as
closely as possible. To this end, even though we have mocked in-memory
providers for the tests, we wrap them in gRPC servers and the test host
communicates with them over gRPC, just as the engine would. Moreover,
the language host under test communicates with the test host over gRPC,
just as it would with a real engine. Presently, however, there is one
missing piece -- when a language host asks the engine to load a schema,
this call is typically answered by a delegated gRPC to the provider
offering that schema. Presently, the conformance test does not do this,
instead loading the schema using the in-memory provider objects it
already possesses. This commit fixes that to match the "real-world"
behaviour, having the test host delegate to its already-gRPC-wrapped
providers instead. To support this new behaviour the test host needs to
learn how to load unversioned providers as well, but this is relatively
simple and shouldn't impact any of the tests we run presently.