fix race conditions in nxadm/tail library (#18044)
Note: This PR is probably best read commit by commit.

The nxadm/tail library we're using has a couple of race conditions, that
may appear when using the automation API. This PR vendors parts of that
library, and pulls in a couple of fixes from upstream (in particular
https://github.com/nxadm/tail/pull/70 and
https://github.com/nxadm/tail/pull/71) to address these race conditions.
This also means we can get rid of the hacky workaround we have for these
upstream issues.

fixes https://github.com/pulumi/pulumi/issues/15235

---------

Co-authored-by: Paul Holzinger <pholzing@redhat.com>