Fix hang on error when using uv on Windows (#18054)
When using uv we were running Python as `uv run python`, since that
conveniently
handles the virtualenv for us. This results in a process tree of
`python-language-runtime -> uv -> python`. This is problematic because
on error
we kill the plugin and its children, but not the children of the
children. On
macOS and Linux, when uv is killed, it kills its children, so we have no
problem
here. On Windows however, it does not, and we end up with an orphaned
Python
process that's busy-waiting in the eventloop and never exits.

To avoid this issue, we run Python the same way as we do when using
Poetry or Pip.

Fixes https://github.com/pulumi/pulumi/issues/17877