Remove docsgen from `pulumi/pulumi` (#18053)
Documentation generation, or "docsgen" is the process by which we
generate HTML documentation from Pulumi schema for various provider
packages. This documentation is hosted as part of the Pulumi Registry,
whose backing code and implementation lives in pulumi/registry. Since
the registry is the only consumer of docsgen, and due to docsgen's
presence in this repository leading to a number of undesirable circular
dependencies, https://github.com/pulumi/registry/pull/6217 copied this
code into the registry's repository directly. This commit thus completes
the move by removing `pulumi/pulumi`'s copy. As part of this we lose the
dependency on `pulumi-yaml`, which is great news. The dependency on
`pulumi-java` remains but should be hopefully easy to remove in a future
change.