don't make backups if the object doesn't exist
We're currently trying to backup files unconditonally in this
function.  However the file doesn't necessarily exist when we call
it (by design).

In the future we want to introduce retries for at least the
bucket.Delete function, so calling that on a file that doesn't exist
would slow down the process.  It really only makes sense to create a
backup if the file does actually exist, so do that.  If we get an
error back from the exists function we still go ahead and try to
create the backup anyway.