Bulk out tests for hydrating registered resources
When a remote component (that is, a resource implemented and returned by
a provider `Construct` method) returns a resource as one of its outputs,
consuming code must "hydrate" that resource using the
`pulumi:pulumi:getResource` `Invoke`, which accepts a URN and produces a
hydrated piece of resource state. Presently, the lifecycle test suite
includes a case exercising this behaviour, but it is not comprehensive.
Specifically, it only checks that a *registered* (i.e. not *read*, by
e.g. a `.get` method) resource can be hydrated. Moreover, it only checks
that such hydration works in the calling program, and not e.g. by
*other* components' `Construct` methods later on. This change renames
and cleans up this case in light of this gap and adds a second test that
checks registered resources in other `Construct` methods. Read resources
will come in a further set of changes.