fix race condition in tail better
In https://github.com/pulumi/pulumi/pull/18044 we vendored the
nxadm/tail library and pulled in a couple of changes from upstream to
avoid race conditions.

However the changes from upstream were not sufficient. Namely,
one of the changes was to try and avoid exiting early if tail is meant
to stop at the next EOF.  It did so by setting the earlyExitChan to
`nil` if tail was requested to shut down with a stopAtEOF error.
However it is possible that stopAtEOF is written to the earlyExitChan
after we checked `tail.Err()`, but before we get to the select.

In that case we can still get an early exit on EOF, which we don't
want.  Move the check for the stopAtEOF error into the earlyExitChan
case. At this point tail.Err() is guaranteed to be set correctly, if
the user requested a stop at eof.