Bulk out tests for hydrating registered resources (#18069)
When a remote component (that is, a resource implemented and returned by
a provider `Construct` method) returns a resource as one of its outputs,
consuming code must "hydrate" that resource using the
`pulumi:pulumi:getResource` `Invoke`, which accepts a URN and produces a
hydrated piece of resource state. Presently, the lifecycle test suite
includes a case exercising this behaviour, but it is not comprehensive.
Specifically, it only checks that a *registered* (i.e. not *read*, by
e.g. a `.get` method) resource can be hydrated. Moreover, it only checks
that such hydration works in the calling program, and not e.g. by
*other* components' `Construct` methods later on. This work makes the
following changes to the lifecycle test suite:

* Splitting out existing tests for remote components/implementations of
the provider `Construct` and `Call` methods into a `component_test`
module.
* Renaming and cleaning up the existing case in light of the gaps
described above.
* Adding a second test that checks registered resources in other
`Construct` methods.

Read resources will come in a further set of changes.

Part of fixing #17515