unflake unhandledInterrupt test (#18075)
It looks like python is occasionally slow in CI, so allow some extra
time for it to shut down to hopefully fix the flake.

Also make the "Took too long to kill child process" case take a little
longer, so we can actually observe what happens in TerminateProcessGroup
and the following asserts, rather than shutting the test down. The test
is still correct, as if we're hitting the 400ms timeout the first assert
in the goroutine will trigger.

Fixes https://github.com/pulumi/pulumi/issues/18074