Fix infinite looping of calls of pulumi program when missing inline program input
described by Fraser here: https://github.com/pulumi/pulumi/issues/15877#issuecomment-2076785682

In summary the program uses the current directory as the project
directory to run instead of an inline program when one is not specified.
This causes pulumi up to be invoked in the current directory (again),
potentially causing an infinite loop.

Fixes #15877