Ignore DependsOn for direct form invokes instead of raising an error
When we introduced DependsOn for output form invokes in Go, we made it an error to pass this option to direct form invokes. Unfortunately that is not backwards compatible, and users can pass the same set of options to resources and invokes. Previously this worked without issue, but now this can result in an error.

Notably, the kubernetes SDK has resources that internally call an invoke with its resource options.

The solution s to to ignore the DependsOn in this case, instead of making it an error. This is what we did for Python and Typescript, [for example](https://github.com/pulumi/pulumi/blob/27cf193189a394a6c1f68583149c90fd4bd512e8/sdk/nodejs/runtime/invoke.ts#L88-L92).

Fixes https://github.com/pulumi/pulumi/issues/18087