Gracefully handle errors that don't implement toString (#18080)
In JavaScript you can throw *anything*, not just errors. Our error
handler was falling back to `”” + err` to stringify the error, however
if that object does not implement `toString()`, the error handler itself
throws with a type error.

We can handle this more gracefully using `util.inspect`, which does not
crash, unless the error is very naughty and overrides the
`util.inspect.custom` hook and throws in that.

Fixes https://github.com/pulumi/pulumi/issues/18068