Ignore DependsOn for direct form invokes instead of raising an error (#18089)
When we introduced DependsOn for output form invokes in Go, we made it
an error to pass this option to direct form invokes. Unfortunately that
is not backwards compatible, and users for example pass the same set of
options to resources and invokes. Previously this worked without issue,
but now this can result in an error.

This also occurs in our SDKs. The Kubernetes SDK has resources that
internally call an invoke with the options of the resource.

The solution is to ignore DependsOn in this case, instead of making it
an error. This is what we did for Python and Typescript, [for
example](https://github.com/pulumi/pulumi/blob/27cf193189a394a6c1f68583149c90fd4bd512e8/sdk/nodejs/runtime/invoke.ts#L88-L92).

Fixes https://github.com/pulumi/pulumi/issues/18087