return a meaningful error when getting a zip template fails (#18091)
When trying to download a ZIP (e.g. for an AI response in pulumi new),
we currently don't check the response code, and just feed whatever we
are getting from the server into the zip parser. The AI at the same time
response with a 500 if the program it generated is not valid, which is
always a possibility with LLMs.

This results in confusing error messages when the AI fails to generate a
correct program: "error: failed to retrieve zip archive: zip: not a
valid zip file".

We can do better here, and check the error code. Helpfully Pulumi AI
also tells us what is wrong in the response body, so we can include that
in the error message.

This results in an error message like (in a failing example I tried):

```
error: failed to retrieve zip archive: failed to download template: 500 Internal Server Error
Missing , or : between flow sequence items at line 30, column 20:

      subnetIDs: [${subnet1.id}]
                   ^
```

This can help the user refine the prompt, and gives them at least a bit
more information on why the AI generation failed.

Fixes https://github.com/pulumi/pulumi/issues/15644