clarify usage string for --expect-no-changes
When using --expect-no-changes in pulumi up and pulumi refresh, the
CLI applies the changes first, and then checks if there were any
later, only erroring out at that point, after having applied the
changes.

This is currently not very clear from either the name of the flag, nor
the usage string.  Improve the usage string to make this clearer.

We could  potentially also "rename" this flag, by hiding the one with
the old name, and introducing a new one doing the same thing, which
would be a backwards compatible change.  I couldn't come up with a
better name for the flag though (error-on-changes is the best I could
come up with, but I don't think it conveys this any better).