Fix reading pulumi-plugin.json when using the uv toolchain (#18107)
Use `pip` instead of `uv pip` to read the package list when determing
pulumi plugins. `uv pip` does not support the `-v` flag, which is
required to retrive the package location with `pip`, see
https://github.com/astral-sh/uv/issues/9838

This requires us to add `pip` as a direct dependency of pulumi.

Fixes https://github.com/pulumi/pulumi/issues/18023