Switch to pyproject.toml + uv (#18081)
Switch from setup.py to pyproject.toml, and from pip+venv to uv.

This gets rid of "build" folder `env` for the Python SDK. Python does
not have to be built, and the SDK can be installed directly from
`sdk/python`, with or without the `--editable` flag. Previously having a
"build" output helped with the SDK version, but this was previously
addressed when we added the version in the source in
https://github.com/pulumi/pulumi/pull/15713.

In a follow up PR we can move ruff and mypy config into
`pyproject.toml`.