Fix prerelease version
With https://github.com/pulumi/pulumi/pull/18081 we changed it so that
the version of the python package is derived from _version.py. However
this file holds the SDK version, which is different from the package
version. Unlink the two versions again to restore the previous
behaviour.

Both versions are bumped via `scripts/set-version.py` when we prepare a
new release. The package version (in pyproject.toml) is additionally
bumped for each pre-release from the master branch.