Fix prerelease version (#18125)
With https://github.com/pulumi/pulumi/pull/18081 we changed it so that
the version of the python package is derived from `_version.py`. However
this file holds the SDK version, which is sometimes different from the
package version. Unlink the two versions again to restore the previous
behaviour.

Both versions are bumped via `scripts/set-version.py` when we prepare a
new release. The package version (in pyproject.toml) is additionally
bumped in CI for each build (and published to PyPI for each merge to
master).