make python auto api tests work with DIY backend
Similar to the nodejs automation API tests in
https://github.com/pulumi/pulumi/pull/17393, we also want to be able
to run the python tests using the DIY backend, rather than just the
cloud backend.  The only changes necessary here are making sure we use
the right organization name (which is always "organization" for the
DIY backend), and making sure the stacks are removed forcefully (for
some reason they seem to think they have resources, even if the up
fails with a compilation error).

When PULUMI_ACCESS_TOKEN is set, the tests will still run with the
cloud backend.