auto/go: interrupt programs gracefully on context cancel (#18124)
When the context is canceled, we currently send a SIGKILL signal to the
process, killing it immediately. This means Pulumi doesn't have time to
clean up after itself as it does on the console when a user presses
Ctrl+C. We can fix that in the "Cancel" handler of the process, by
sending a SIGINT first (and setting WaitDelay, so the process eventually
gets killed if it's not responding anymore).

Fixes https://github.com/pulumi/pulumi/issues/13160