Avoid calling invokes with dependencies on unknwon resources
DependsOn for resources is an ordering constraint for register resource calls. If a resource R1 depends on a resource R2, the register resource call for R2 will happen after R1. This is ensured by awaiting the URN for each resource dependency before calling register resource.

For invokes, this causes a problem when running under preview. During preview, register resource immediately returns with the URN, however this does not tell us if the resource "exists".

Instead of waiting for the dependency's URN, we wait for the ID. This tells us that wether a physical resource exists (if the state does not require a refresh), and we can avoid calling the invoke when it is unknown.