Add a conformance test for program-hydrated resource references (#18139)
This change adds the `l2-component-program-resource-ref` conformance
test, which tests that a language supports generating SDKs and programs
that are capable of handling resource references returned from remote
components and hydrating them in the program so that their outputs can
be used later on (in this case, as inputs to subsequent custom
resources). Subsequent changes will add tests for in-component
hydration, as well as other component-related features such as method
invocations (`Call`). The idea is that this will help us shore up
component-related functionality, particularly with respect to suspect
gaps in `pulumi-java`.

This test already passes for NodeJS and Python, but fails in Go due to a
bug in program generation. It is therefore skipped for Go with the bug
being tracked as #18140.

Working towards closing pulumi/pulumi-java#262