Add a conformance test for component-hydrated resource references
This change adds the `l2-component-component-resource-ref` conformance
test, which tests that a language supports generating SDKs and programs
that are capable of handling resource references returned from remote
components and hydrating them in other components so that their outputs
can be used later on (in this case, as inputs to subsequent custom
resources).

This test already passes for NodeJS and Python, but fails in Go due to a
bug in program generation. It is therefore skipped for Go with the bug
being tracked as #18140.

Working towards closing pulumi/pulumi-java#262