go automation api support for colorize output
Currently, the Options struct has a field for color, but there is not
functions exposed to set this field in go. This change adds the ability
to set the color setting for output. optdestroy, optpreview, optrefresh,
and optup all have a new function to set the color setting.

To use the new color setting, you can set the color field in the Options
struct like the following:

```go
	_, err = s.Up(ctx, stdoutStreamer, optup.Color("always"))
	if err != nil {
		fmt.Printf("Failed to update stack: %v\n\n", err)
		os.Exit(1)
	}
```