Adds support for Copilot templates to `pulumi new` (#16825)
Copilot templates require an access token to look up the conversation
data and extract the code

Here we try to download any template as usual, but then add the
following logic to the response checking:
- *If* a `403` is returned from what we identify as a pulumi-service
backend (via `X-Pulumi-Request-ID` header, which has been present for
the last 6 years)
- *Then* we re-request w/ the auth token. If there is no token present
then the login flow is automatically initiated.

This does result in always raising a 403 on the server every time we
request a copilot template.

If we want to avoid these 403s we could check the domain against known
pulumi backends in the CLI first, at the cost of intro'ing another code
path.

## To test

- Go through the Copilot flow, e.g "Create an s3 bucket please"
- Click "Deploy with Pulumi"
- Choose CLI, click next
- Check out this branch
- Copy most of the command but first: `cd ~/pulumi/pulumi/pkg` and then
`go run ./cmd/pulumi new $ARGS_FROM_PULUMI_CLOUD_UI`

It should succeed