Add more documentation on PCL (#18193)
PCL, or Pulumi Configuration Language, is the core intermediate language
that allows several parts of Pulumi to work in a language-agnostic
manner. This change documents some of the processes and functionality
that we expose for working with PCL programs, incorporating some
existing documentation on extensions we make to e.g. HCL's type system.